--- eyamlpath/__init__.py
@@ -1,1 +1,0 @@
-from .eyamlhelpers import EYAMLHelpers
