+++ yamlpath/enums/__init__.py
@@ -1,0 +1,3 @@
+from yamlpath.enums.yamlvalueformats import YAMLValueFormats
+from yamlpath.enums.pathsegmenttypes import PathSegmentTypes
+from yamlpath.enums.pathsearchmethods import PathSearchMethods
