+++ yamlpath/enums/pathsearchmethods.py
@@ -1,54 +1,5 @@
 from enum import Enum, auto
 
-class YAMLValueFormats(Enum):
-    """Supported representation formats for YAML values."""
-    BARE = auto()
-    BOOLEAN = auto()
-    DEFAULT = auto()
-    DQUOTE = auto()
-    FLOAT = auto()
-    FOLDED = auto()
-    INT = auto()
-    LITERAL = auto()
-    SQUOTE = auto()
-
-    @staticmethod
-    def get_names():
-        """Returns all entry names for this enumeration.
-
-        Positional Parameters:  N/A
-
-        Returns:  (list) Upper-case names from this enumeration
-
-        Raises:  N/A
-        """
-        return [entry.name.upper() for entry in YAMLValueFormats]
-
-    @staticmethod
-    def from_str(name):
-        """Converts a string value to a value of this enumeration, if valid.
-
-        Positional Parameters:
-          1. name (str) The name to convert
-
-        Returns:  (YAMLValueFormats) the converted enumeration value
-
-        Raises:
-          NameError when name doesn't match any enumeration values.
-        """
-        check = str(name).upper()
-        if check in YAMLValueFormats.get_names():
-            return YAMLValueFormats[check]
-        else:
-            raise NameError("YAMLValueFormats has no such item, " + check)
-
-class PathSegmentTypes(Enum):
-    """Supported YAML Path segments"""
-    ANCHOR = auto()
-    INDEX = auto()
-    KEY = auto()
-    SEARCH = auto()
-
 class PathSearchMethods(Enum):
     """Supported methods for search YAML Path segments"""
     CONTAINS = auto()
