+++ yamlpath/enums/pathsegmenttypes.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+class PathSegmentTypes(Enum):
+    """Supported YAML Path segments"""
+    ANCHOR = auto()
+    INDEX = auto()
+    KEY = auto()
+    SEARCH = auto()
