+++ yamlpath/exceptions/__init__.py
@@ -1,0 +1,1 @@
+from yamlpath.exceptions.yamlpath import YAMLPathException
