+++ yamlpath/eyaml/__init__.py
@@ -1,0 +1,1 @@
+from .eyamlpath import EYAMLPath
